refactor: 어드민 페이지 분리 (props, 유틸 포맷함수 등 미분리)
